==> 1_PEMr_prepare_project.Rmd <==
---
title: "ML_PEM full workflow"
subtitle: "Step through all functions"
author: "W. H.MacKenzie"
date: "2023-06-19"
output: rmarkdown::html_vignette
vignette: >
  %\VignetteIndexEntry{Set up new study area}
  %\VignetteEncoding{UTF-8}
  %\VignetteEngine{knitr::rmarkdown}
editor_options:
  chunk_output_type: console
---

```{r, include = FALSE}
knitr::opts_chunk$set(  collapse = TRUE,  comment = "#>")
```

```{r setup, message=FALSE, warning=FALSE}
# remotes::install_github('bcgov/PEMr', build_vignettes = FALSE)

library(PEMr)
run_pemr()
library(PEMprepr)
library(ggplot2)
library(sf)
library(terra)
library(PEMsamplr)
require(clhs)
require(data.table)

devtools::load_all("D:\\GitHub\\PEMprepr")
devtools::load_all("D:\\GitHub\\PEMsamplr")
```

`PEMr`, short for _Predictive Ecosystem Mapping in R is a set of packages for developing a machine learning PEM following the process of MacKenzie et al. 
See [BC-BEC](https://www.for.gov.bc.ca/HRE/becweb/index.html)
and [`PEMr`](https://github.com/bcgov/PEMr) on _GitHub_for details about the BC BEC classification system and, the PEM Project.

In the example below it is assumed that you are running the following within a new R project.
## Project Set up. 
1) Set up a new R studio project for your PEM modelling. 
2) For each study area a folder structure based on your project/AOI name is constructed under this Rproject folder 

# Project folder setup
```{r create project folders}
project_name <- "WetzinkwaProcess_AOI"
fid <- setup_folders(project_name)

```

##Add base data to the folders

Copy an aoi.gpkg (project boundary) into the Project /00_raw_inputs/10_vector/ folder.
```{r}

aoi_snap <- function(aoi, method=c("expand","shrink"), buffer=0){
  ## testing
  # setwd("e:/workspace/2019/PEM_2020/PEMWorkFlow/")
  #aoi <- sf::st_read("../data/Block_aoi.gpkg")
  bb <- sf::st_bbox(aoi)

  ## Function
  print("initial extent is:")
  print(bb)

  if (buffer>0 & method == "expand") {
    ## Generate expanded bbox -- expands to neared 100m
    xmin <- floor((bb$xmin - buffer) / 100)*100
    xmax <- ceiling((bb["xmax" ]+ buffer) / 100) * 100
    ymin <- floor((bb$ymin - buffer) / 100)*100
    ymax <- ceiling((bb["ymax"] + buffer) / 100) * 100


} else if (method == "expand") {
    ## Generate expanded bbox -- expands to neared 100m
    xmin <- floor(bb$xmin / 100)*100
    xmax <- ceiling(bb["xmax"] / 100) * 100
    ymin <- floor(bb$ymin / 100)*100
    ymax <- ceiling(bb["ymax"] / 100) * 100

  } else if (method == "shrink") {

    xmin <- ceiling(bb$xmin / 100)*100
    xmax <- floor(bb["xmax"] / 100) * 100
    ymin <- ceiling(bb$ymin / 100)*100
    ymax <- floor(bb["ymax"] / 100) * 100

  }

  box <- matrix(c(xmin, ymin, xmin, ymax, xmax, ymax, xmax, ymin, xmin, ymin), ncol = 2, byrow = TRUE)
  box <- sf::st_polygon(list(box))
  box <- sf::st_sfc(box, crs=sf::st_crs(aoi))
  box <- sf::st_as_sf(box)

  ## Report and Return
  print("Extent is:")
  print(sf::st_bbox(box))
  return(box)

}

```

```{r function to create aoi bounding box}

create_aoi <- function(aoi_dir = file.path(fid$shape_dir_0010[2]), 
                       out_dir = file.path(fid$shape_dir_1010[1]), 
                       aoi_vec = "aoi.gpkg",
                       buffer=1000){       ## input raster
aoi <- st_read(file.path(aoi_dir, aoi_vec))
aoi_bb <- aoi_snap(aoi, method = "expand")
aoi <- st_read(file.path(aoi_dir, aoi_vec))
#aoi_ls <- aoi_snap_watershed(aoi, "expand")
aoi_buffered <- aoi_snap(aoi, method = "expand", buffer = buffer)
sf::st_write(aoi, file.path(out_dir,"aoi.gpkg"), append = FALSE)
sf::st_write(aoi_bb, file.path(out_dir,"aoi_snapped.gpkg"), append = FALSE)
sf::st_write(aoi_buffered, file.path(out_dir,"aoi_buffered.gpkg"), append = FALSE)
#sf::st_write(aoi_ls, file.path(out_dir,"aoi_watershed.gpkg"), append = FALSE)
}

```


```{r create aoi bounding box}
create_aoi(buffer = 1000)
```

##Downloads AOI vector data from BC Data catalogue
A standard set of Vector data is used for subsequent sample planning and is based on the `aoi`.  This script collects numerous vector layers from the [BC Data Catalogue](https://catalogue.data.gov.bc.ca) making use of the [`bcdata`](https://github.com/bcgov/bcdata) package. Files are saved in _geopackage_ format.
Other option would be to simply point to an raw folder where files have been copied and snap all those layers

```{r, eval=FALSE}
###----changed default parameters for file locations and added output list
#devtools::load_all("D:\\GitHub\\PEMprepr")
create_base_vectors()
v <- list.files(path = fid$shape_dir_0010[1], pattern = ".gpkg", recursive = TRUE)
v
```


## Review data and make anyfixes
Once the vector data is reviewed and edited where appropriate we can save a final version in the "clean input data folder". 
```{r copy clean vectors function}
# copy_clean_vectors <- function(origindir = fid$shape_dir_0010[1], 
#                        targetdir = fid$shape_dir_1010[1], 
#                        aoi_vec = "aoi.gpkg"){       ## input raster
# filestocopy <- list.files(path = origindir, pattern = ".gpkg",
#     recursive = TRUE)
# lapply(filestocopy, function(x) file.copy(paste(origindir, x, sep = "/"),
#     paste(targetdir, x, sep = "/"), recursive = FALSE, copy.mode = TRUE))
# }
```


```{r copy reviewed layers to clean folders}
copy_clean_vectors()
```

## Create raster templates from AOI

The snapped `aoi` is used to create a raster template all generated covariates created to match selected template.

```{r create raster templates function}

create_template <- function(aoi_bb = file.path(fid$shape_dir_1010[1],
                                               "aoi_snapped.gpkg"), 
                            res,
                            outpath = fid$cov_dir_1020[2],
                            filename = "template.tif"){

  res_name = paste0(res, 'm')
  res_dir = file.path(outpath,res_name)
  #res$exists <- dir.exists(res_dir)
  if(dir.exists(res_dir) == FALSE) {

    print("creating folders")
    dir.create(res_dir, recursive = TRUE)
    print("folder structure created")
    
  } else {

    print(paste("The", as.character(res_dir), "folder already exists - returning folder names."))
    aoi_bb <- terra::vect(aoi_bb)
    template <- terra::rast(aoi_bb, resolution = res)
      terra::values(template) <- 0
    terra::writeRaster(template, file.path(outpath, res_name, filename), overwrite = TRUE)
  return(template)

    }
}

```


```{r create raster templates}
create_template(res = 5)
create_template(res = 25)
```

# Generation of raster covariates

Here we generate the terrain derived features for modelling.  This is powered by [SAGA-GIS](https://saga-gis.sourceforge.io/en/index.html).  
Covariates can be generated at multiple resolutions. The standard modelling for ML-PEM is at 5m scale.

1. Copy a dem.tif raster layer into the ./00_raw_inputs/20_dem/lidar folder.
2. Align the dem to the template.  resample the `dtm` to the desired resolution.
3. create the predictor covariates

## Import LiDAR DEM

```{r read in local LiDAR DEM}
dem <- terra::rast(file.path(fid$lidar_dir_00[2], "dem.tif"))
r5 <-  terra::rast(file.path(fid$cov_dir_1020, "/5m/template.tif"))
dem <- terra::project(dem, r5)
terra::writeRaster(dem, file.path(fid$cov_dir_1020[2], "/5m/dem.tif"), overwrite = TRUE)
plot(dem)
## or alternate get external TRIM data for DEM

```


```{r read in Trim DEM from bcmaps}
library(bcmaps)

get_trim_dem <- function(aoi_bb = file.path(fid$shape_dir_1010[2],"aoi_snapped.gpkg"),
                         template = file.path(fid$cov_dir_1020[2],res_name, "template.tif"),
                         res,
                         filename = "dem.tif",
                         out_dir = fid$cov_dir_1020[2]){
aoi <- st_read(aoi_bb)
  trim_raw <- bcmaps::cded_raster(aoi)
trim <- terra::rast(trim_raw)
res_name = paste0(res, 'm')
template_file <- template
template <- terra::rast(template_file)
trim <- terra::project(trim, template)
terra::writeRaster(trim, file.path(out_dir, res_name, filename), overwrite = TRUE)
}


```

```{r get TRIM dem at scales}
#get_trim_dem(res = 5)
get_trim_dem(res = 25)
```

## Create the covariates

```{r, eval=FALSE}

layer_options <- c("sinksfilled", "sinkroute", "dem_preproc", "slope_aspect_curve",
               "tcatchment", "tca", "scatchment", "twi", "channelsnetwork",
               "overlandflow", "overlandflow2", "multiresflatness", "multiresflatness2",
               "multiresflatness3", "tri", "convergence", "openness",
               "dah", "tpi", "ridgevalley", "mrn", "flowaccumulation",
               "slopelength", "flowaccumulation2", "flowaccumulation3",
               "flowpathlength", "flowpathlength2", "flowpathlength3", "lsfactor",
               "solarrad", "convexity", "vertdistance", "tci_low",
               "swi", "windexp", "texture", "protection", "vrm",
               "mbi", "mscale_tpi", "relposition", "slopecurvatures",
               "steepestslope", "upslopearea")


  # run a test covariate
devtools::load_all("D:\\GitHub\\PEMprepr")

####add these a default parameters to create_covariates
sagapath = "C:/saga-9.0.2_x64/"
dem = file.path(fid$cov_dir_1020[2], "/5m/dem.tif")
create_covariates(dtm = dem,        ## raster created above
                  SAGApath = sagapath, ## Where SAGA GIS is installed
                  output = fid$cov_dir_1020[2], ## from the setup_folders above
                  layers = "sinksfilled")        ## test one is working 

# run all covariates 
create_covariates(dtm = dem,           ## raster created above
                  SAGApath = sagapath, ## Where SAGA GIS is installed
                  output = fid$cov_dir_1020[2], ## from the setup_folders above
                  layers = c( "all"))       ## use all or one of the above

##check the list of covariates. Notes these are saved in sdat format.

l <- list.files(path = fid$cov_dir_1020[2], pattern = ".sdat$",
                recursive = TRUE)

l
```

Create a bec raster to match the template sizes 

```{r create rast from vector using template}
###create function
# rasterize_to_template <- function(covar = "bec", field = "MAP_LABEL", template = r5){
#   if (class(covar)[1] == "sf") {
#     aoi <- terra::vect(aoi)
#   }
# covar_sf <- sf::st_read(file.path(fid$shape_dir_1010[2], paste0(covar, ".gpkg"))) %>%   sf::st_cast(., "MULTIPOLYGON")
# covar_code <- covar_sf %>% st_drop_geometry()  %>% dplyr::select(all_of(field)) %>%
#     unique() 
# covar_code <- covar_code %>% 
#     mutate(covar_unique_code = seq(1, length(covar_code[1]),1))
# covar_sf <- dplyr::left_join(covar_sf, covar_code)
# covar_vect <- terra::vect(covar_sf)
# covar_rast <- terra::rasterize(covar_vect, template)
# return(covar_rast)
# }
```


```{r}
#bec_ras5 <- rasterize_to_template()
bec_ras5 <- create_bgc_template()
terra::writeRaster(bec_ras5, file.path(fid$cov_dir_1020[2], "5m", "bec.tif"), overwrite = TRUE)
```
